Перенесла в темплейт часть разметки из формы редактирования (в виде функций), удалила ненужные вьюшки
